test: harden synthetic idiffir e2e fixture and assertions
